Rename tao classes and packages.
